attempt to read-up script dialog into some kind of struct
the idea is that each line will contain a word count. the user will
get to select how many words per paragraph. we'll randomly select
lines of dialog until we reach this man words (roughly). we'll also
randomly select each line of dialog from more than one script. but
right now we only have one.

println is required to know that this worked since the sequence that
is returned is lazy.